pyproject.toml
[project]
name = "api-reservations"
version = "0.1.0"
description = "Reservations backend (FastAPI, MySQL, Stripe, suppliers)"
requires-python = ">=3.12,<3.14"
dependencies = [
  "fastapi>=0.110.0,<1.0.0",
  "uvicorn[standard]>=0.27.0,<0.29.0",
  "pydantic>=2.6.0,<3.0.0",
  "email-validator>=2.1.0,<2.2.0",
  "pydantic-settings>=2.2.0,<3.0.0",
  "sqlalchemy[asyncio]>=2.0.25,<2.1.0",
  "aiomysql>=0.2.0,<0.3.0",
  "stripe>=8.0.0,<10.0.0",
  "httpx>=0.27.0,<0.29.0",
  "google-generativeai>=0.8.0,<1.0.0",
  "aiosqlite>=0.22.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "httpx>=0.27.0,<0.28.0",
  "ruff>=0.3.0,<0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "pyproject.toml"]
